Fix mysql_upgrade possible race condition
On deb packages of MariaDB, mysql_upgrade is ran with debian-start
script as systemd post-run step. This means, that we don't really have
control when it's launched and we need to wait until process
is finished before proceeding with it's setup.
In the meanwhile redhat is packaged differently and mysql_upgrade
must be run manually.

Based on that we change ordering and verifying explicitly that
upgrade has finished.

Change-Id: Ib0f410cfcdcc416b58b03febab9c55dc7982de38