Move slow_query_log_file to location handled by logrotation
Currently slow_query_log_file is not set and the default of `host_name-slow.log` is applied.
This causes an ever growing slow log to fill up `/var/lib/mysql` and which is never rotated.

By placing this file at `/var/log/mysql/mariadb-slow.log` it will be rotated by the bundled
logrotate config of the `mysql-server` package.

Change-Id: Ib66eb5c6bdf94b6c6f4461a7f6e339c1000e0afc