Bump mariadb version.
This bump will fix:
https://nvd.nist.gov/vuln/detail/CVE-2022-27376
https://nvd.nist.gov/vuln/detail/CVE-2022-27377
https://nvd.nist.gov/vuln/detail/CVE-2022-27380

We also modify workaround applied for 10.5 upgrade wrt to bug [1]
as it has been added extra tools to help with checking state of upgrade.

New flag --check-if-upgrade-is-needed is checking if any upgrade is
already running and waits until it's finished.
It exits with rc 0 if upgrade is required and 1 if not.

If upgrade is required, we fall into rescue and perform upgrade.

[1] https://jira.mariadb.org/browse/MDEV-27068

Related-Bug: #1979728

Change-Id: I90286b01b5578df04447155c69c92f823388185a