Define backup randomized delay in defaults
Omit can not be used in timer options, since this is simple mapping
that is passed to the unit file. With that, omit is resolved to a
randomly named omit_place_holder  that ends up in a template.

Se we define a delay to 0, which is default systemd behaviour [1]

[1] https://www.freedesktop.org/software/systemd/man/systemd.timer.html#RandomizedDelaySec=

Change-Id: Ib242e66cfb4a24b7e93144e382e50f124015e3bf
(cherry picked from commit 92b5711b94d734d77665875bb99deb09cda72b25)