Fix ignored database directories configuration
Confusingly, the variable ignore_db_dirs is set by passing it
multiple times in the configuration file, once per directory.
It is then read as a comma separated list, but cannot be set
in this way.
https://mariadb.com/kb/en/server-system-variables/#ignore_db_dirs

Without this, the mariadb-upgrade script can fail as it attempts
to process invalid databases.

Change-Id: Ie997393935e04e127893643e4c72d7af07e993ff