Do not verify MariaDB version when installed from distro repos
MariaDB which is installed from native distro repositories will not
match our expected version. We also should not really verify it,
as it's not up to the role to manage the upgrade cycle for the
database in case of distro source.

Change-Id: I9c30c8c61bb4f1d015baf997955a66adc4df01c4
(cherry picked from commit ce844bc4da2f2479374f17564c92c2fd60b4a4c7)